workload_app/requirements.txt
diagrams
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-containerinstance
azure-mgmt-network
azure-mgmt-storage

dash
networkx
plotly
dash-bootstrap-components
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
